fix: Use Buster-based Ruby for builder image
The `ruby:2.6.8` image tag was changed some hours ago by the Ruby team
to be based on Debian 11 `bullseye` instead of Debian 9 `stretch`,
which caused native dependency errors from `ffi` gem.
This changes the base image for the builder image to `ruby:2.6.8-buster`
to avoid the ffi error. It also matches the Fluentd image,
which is also based on `buster`.